pss: refactor. Extracted outbox and outboxMsg into outbox.go
